chore(pkg/client): move heartbeater.go to pkg/client
